requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arxiv==1.4.7
asgiref==3.6.0
astroid==2.15.4
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==40.0.2
dataclasses-json==0.5.7
defusedxml==0.7.1
dill==0.3.6
dj-database-url==1.2.0
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.52.0
django-simpleui==2023.3.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.2.2
faiss-cpu==1.7.4
feedparser==6.0.10
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
isort==5.12.0
langchain==0.0.159
lazy-object-proxy==1.9.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
platformdirs==3.5.0
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pylint==2.17.3
PyMuPDF==1.22.2
pypdf==3.8.1
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.27.1
requests-oauthlib==1.3.1
sgmllib3k==1.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
sqlparse==0.4.4
tenacity==8.2.2
tiktoken==0.4.0
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
yarl==1.9.2
docx2txt==0.8
